feat: Adding nosql database (redis)
Currently, only redis is added, but more databases will be added in the next version and updates.

The makefile tool was written for ease of use.

Docker files for dockerizing and building a container for the project were completely created.

A unit test was added to connect to redis and execute one or more queries